requirements.txt
anthropic>=0.40.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
ta>=0.11.0
requests>=2.31.0
schedule>=1.2.0
